DTFS2-4677 - fix mock-data-loader delete tfm deal issue with objectId
