+++ dtfs-central-api/src/v1/controllers/tfm/deal/tfm-delete-deal.controller.js
@@ -1,4 +1,3 @@
-const { ObjectId } = require('mongodb');
 const { findOneDeal } = require('./tfm-get-deal.controller');
 const db = require('../../../../drivers/db-client');
 
@@ -7,7 +6,7 @@ exports.deleteDeal = async (req, res) => {
     if (deal) {
       const collection = await db.getCollection('tfm-deals');
       const facilitiesCollection = await db.getCollection('tfm-facilities');
-      const status = await collection.deleteOne({ _id: ObjectId(req.params.id) });
+      const status = await collection.deleteOne({ _id: req.params.id });
 
       await facilitiesCollection.deleteMany({ 'facilitySnapshot.associatedDealId': { $eq: deal._id } });
       await facilitiesCollection.deleteMany({ 'facilitySnapshot.applicationId': { $eq: deal._id } });
