+++ utils/mock-data-loader/centralApi.js
@@ -12,20 +12,6 @@ const getDeal = async (dealId) => {
   return response.data;
 };
 
-const deleteDeal = async (deal, token) => {
-  const response = await axios({
-    method: 'delete',
-    headers: {
-      'Content-Type': 'application/json',
-      Accepts: 'application/json',
-      Authorization: token || '',
-    },
-    url: `${urlRoot}/v1/tfm/deals/${deal._id}`,
-  }).catch(() => { });
-
-  return response && response.data;
-};
-
 const createFacility = async (facility, associatedDealId, user) => {
   const response = await axios({
     method: 'post',
@@ -89,7 +75,6 @@ const deleteFacility = async (facilityId) => {
 
 module.exports = {
   getDeal,
-  deleteDeal,
   createFacility,
   updateFacility,
   listFacilities,
